+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: mde-krui <mde-krui@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/10/02 17:17:47 by mde-krui      #+#    #+#                 */
-/*   Updated: 2023/10/09 15:03:36 by mde-krui      ########   odam.nl         */
+/*   Updated: 2023/10/11 12:01:48 by mde-krui      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (src[i] && i + 1 < size)
